+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 
 * Add `ruby-3.0` to CI
+* Add `codecov` branch coverage check in CI
 
 ## 1.1.0 (2020-11-17)