+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 * Add `ruby-3.0` to CI
 * Add `codecov` branch coverage check in CI
 * Add new methods:
-  * green_log()
-  * red_log()
+  * OnlyofficeLoggerHelper.green_log()
+  * OnlyofficeLoggerHelper.red_log()
 
 ## 1.1.0 (2020-11-17)